@@ -1,287 +1,31 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
+  <Target Name="PostBuild" AfterTargets="ComputeFilesToPublish">
+    <Exec Command="npm install"/>
+  </Target>
+
   <ItemGroup>
-    <Content Include="node_modules\bootstrap\dist\css\bootstrap-theme.css" />
-    <Content Include="node_modules\bootstrap\dist\css\bootstrap-theme.min.css" />
-    <Content Include="node_modules\bootstrap\dist\css\bootstrap.css" />
-    <Content Include="node_modules\bootstrap\dist\css\bootstrap.min.css" />
-    <Content Include="node_modules\bootstrap\dist\fonts\glyphicons-halflings-regular.eot" />
-    <Content Include="node_modules\bootstrap\dist\fonts\glyphicons-halflings-regular.ttf" />
-    <Content Include="node_modules\bootstrap\dist\fonts\glyphicons-halflings-regular.woff" />
-    <Content Include="node_modules\bootstrap\fonts\glyphicons-halflings-regular.eot" />
-    <Content Include="node_modules\bootstrap\fonts\glyphicons-halflings-regular.ttf" />
-    <Content Include="node_modules\bootstrap\fonts\glyphicons-halflings-regular.woff" />
-    <Content Include="node_modules\bootstrap\grunt\configBridge.json" />
-    <Content Include="node_modules\bootstrap\package.json" />
-    <Content Include="node_modules\datatables.net-bs\css\dataTables.bootstrap.css" />
-    <Content Include="node_modules\datatables.net-bs\css\dataTables.bootstrap.min.css" />
-    <Content Include="node_modules\datatables.net-bs\License.txt" />
-    <Content Include="node_modules\datatables.net-bs\package.json" />
-    <Content Include="node_modules\datatables.net\License.txt" />
-    <Content Include="node_modules\datatables.net\package.json" />
-    <Content Include="node_modules\jquery\AUTHORS.txt" />
-    <Content Include="node_modules\jquery\bower.json" />
-    <Content Include="node_modules\jquery\external\sizzle\LICENSE.txt" />
-    <Content Include="node_modules\jquery\LICENSE.txt" />
-    <Content Include="node_modules\jquery\package.json" />
+    <Content Include="node_modules/**" CopyToPublishDirectory="PreserveNewest"/>
   </ItemGroup>
 
   <ItemGroup>
-    <None Include="node_modules\bootstrap\CHANGELOG.md" />
-    <None Include="node_modules\bootstrap\dist\css\bootstrap-theme.css.map" />
-    <None Include="node_modules\bootstrap\dist\css\bootstrap-theme.min.css.map" />
-    <None Include="node_modules\bootstrap\dist\css\bootstrap.css.map" />
-    <None Include="node_modules\bootstrap\dist\css\bootstrap.min.css.map" />
-    <None Include="node_modules\bootstrap\dist\fonts\glyphicons-halflings-regular.svg" />
-    <None Include="node_modules\bootstrap\dist\fonts\glyphicons-halflings-regular.woff2" />
-    <None Include="node_modules\bootstrap\dist\js\bootstrap.js" />
-    <None Include="node_modules\bootstrap\dist\js\bootstrap.min.js" />
-    <None Include="node_modules\bootstrap\dist\js\npm.js" />
-    <None Include="node_modules\bootstrap\fonts\glyphicons-halflings-regular.svg" />
-    <None Include="node_modules\bootstrap\fonts\glyphicons-halflings-regular.woff2" />
-    <None Include="node_modules\bootstrap\Gruntfile.js" />
-    <None Include="node_modules\bootstrap\grunt\.jshintrc" />
-    <None Include="node_modules\bootstrap\grunt\.stylelintrc" />
-    <None Include="node_modules\bootstrap\grunt\browsers.js" />
-    <None Include="node_modules\bootstrap\grunt\bs-commonjs-generator.js" />
-    <None Include="node_modules\bootstrap\grunt\bs-glyphicons-data-generator.js" />
-    <None Include="node_modules\bootstrap\grunt\bs-lessdoc-parser.js" />
-    <None Include="node_modules\bootstrap\grunt\bs-raw-files-generator.js" />
-    <None Include="node_modules\bootstrap\grunt\change-version.js" />
-    <None Include="node_modules\bootstrap\grunt\generate-sri.js" />
-    <None Include="node_modules\bootstrap\grunt\karma.conf.js" />
-    <None Include="node_modules\bootstrap\js\affix.js" />
-    <None Include="node_modules\bootstrap\js\alert.js" />
-    <None Include="node_modules\bootstrap\js\button.js" />
-    <None Include="node_modules\bootstrap\js\carousel.js" />
-    <None Include="node_modules\bootstrap\js\collapse.js" />
-    <None Include="node_modules\bootstrap\js\dropdown.js" />
-    <None Include="node_modules\bootstrap\js\modal.js" />
-    <None Include="node_modules\bootstrap\js\popover.js" />
-    <None Include="node_modules\bootstrap\js\scrollspy.js" />
-    <None Include="node_modules\bootstrap\js\tab.js" />
-    <None Include="node_modules\bootstrap\js\tooltip.js" />
-    <None Include="node_modules\bootstrap\js\transition.js" />
-    <None Include="node_modules\bootstrap\less\alerts.less" />
-    <None Include="node_modules\bootstrap\less\badges.less" />
-    <None Include="node_modules\bootstrap\less\bootstrap.less" />
-    <None Include="node_modules\bootstrap\less\breadcrumbs.less" />
-    <None Include="node_modules\bootstrap\less\button-groups.less" />
-    <None Include="node_modules\bootstrap\less\buttons.less" />
-    <None Include="node_modules\bootstrap\less\carousel.less" />
-    <None Include="node_modules\bootstrap\less\close.less" />
-    <None Include="node_modules\bootstrap\less\code.less" />
-    <None Include="node_modules\bootstrap\less\component-animations.less" />
-    <None Include="node_modules\bootstrap\less\dropdowns.less" />
-    <None Include="node_modules\bootstrap\less\forms.less" />
-    <None Include="node_modules\bootstrap\less\glyphicons.less" />
-    <None Include="node_modules\bootstrap\less\grid.less" />
-    <None Include="node_modules\bootstrap\less\input-groups.less" />
-    <None Include="node_modules\bootstrap\less\jumbotron.less" />
-    <None Include="node_modules\bootstrap\less\labels.less" />
-    <None Include="node_modules\bootstrap\less\list-group.less" />
-    <None Include="node_modules\bootstrap\less\media.less" />
-    <None Include="node_modules\bootstrap\less\mixins.less" />
-    <None Include="node_modules\bootstrap\less\mixins\alerts.less" />
-    <None Include="node_modules\bootstrap\less\mixins\background-variant.less" />
-    <None Include="node_modules\bootstrap\less\mixins\border-radius.less" />
-    <None Include="node_modules\bootstrap\less\mixins\buttons.less" />
-    <None Include="node_modules\bootstrap\less\mixins\center-block.less" />
-    <None Include="node_modules\bootstrap\less\mixins\clearfix.less" />
-    <None Include="node_modules\bootstrap\less\mixins\forms.less" />
-    <None Include="node_modules\bootstrap\less\mixins\gradients.less" />
-    <None Include="node_modules\bootstrap\less\mixins\grid-framework.less" />
-    <None Include="node_modules\bootstrap\less\mixins\grid.less" />
-    <None Include="node_modules\bootstrap\less\mixins\hide-text.less" />
-    <None Include="node_modules\bootstrap\less\mixins\image.less" />
-    <None Include="node_modules\bootstrap\less\mixins\labels.less" />
-    <None Include="node_modules\bootstrap\less\mixins\list-group.less" />
-    <None Include="node_modules\bootstrap\less\mixins\nav-divider.less" />
-    <None Include="node_modules\bootstrap\less\mixins\nav-vertical-align.less" />
-    <None Include="node_modules\bootstrap\less\mixins\opacity.less" />
-    <None Include="node_modules\bootstrap\less\mixins\pagination.less" />
-    <None Include="node_modules\bootstrap\less\mixins\panels.less" />
-    <None Include="node_modules\bootstrap\less\mixins\progress-bar.less" />
-    <None Include="node_modules\bootstrap\less\mixins\reset-filter.less" />
-    <None Include="node_modules\bootstrap\less\mixins\reset-text.less" />
-    <None Include="node_modules\bootstrap\less\mixins\resize.less" />
-    <None Include="node_modules\bootstrap\less\mixins\responsive-visibility.less" />
-    <None Include="node_modules\bootstrap\less\mixins\size.less" />
-    <None Include="node_modules\bootstrap\less\mixins\tab-focus.less" />
-    <None Include="node_modules\bootstrap\less\mixins\table-row.less" />
-    <None Include="node_modules\bootstrap\less\mixins\text-emphasis.less" />
-    <None Include="node_modules\bootstrap\less\mixins\text-overflow.less" />
-    <None Include="node_modules\bootstrap\less\mixins\vendor-prefixes.less" />
-    <None Include="node_modules\bootstrap\less\modals.less" />
-    <None Include="node_modules\bootstrap\less\navbar.less" />
-    <None Include="node_modules\bootstrap\less\navs.less" />
-    <None Include="node_modules\bootstrap\less\normalize.less" />
-    <None Include="node_modules\bootstrap\less\pager.less" />
-    <None Include="node_modules\bootstrap\less\pagination.less" />
-    <None Include="node_modules\bootstrap\less\panels.less" />
-    <None Include="node_modules\bootstrap\less\popovers.less" />
-    <None Include="node_modules\bootstrap\less\print.less" />
-    <None Include="node_modules\bootstrap\less\progress-bars.less" />
-    <None Include="node_modules\bootstrap\less\responsive-embed.less" />
-    <None Include="node_modules\bootstrap\less\responsive-utilities.less" />
-    <None Include="node_modules\bootstrap\less\scaffolding.less" />
-    <None Include="node_modules\bootstrap\less\tables.less" />
-    <None Include="node_modules\bootstrap\less\theme.less" />
-    <None Include="node_modules\bootstrap\less\thumbnails.less" />
-    <None Include="node_modules\bootstrap\less\tooltip.less" />
-    <None Include="node_modules\bootstrap\less\type.less" />
-    <None Include="node_modules\bootstrap\less\utilities.less" />
-    <None Include="node_modules\bootstrap\less\variables.less" />
-    <None Include="node_modules\bootstrap\less\wells.less" />
-    <None Include="node_modules\bootstrap\LICENSE" />
-    <None Include="node_modules\bootstrap\README.md" />
-    <None Include="node_modules\datatables.net-bs\js\dataTables.bootstrap.js" />
-    <None Include="node_modules\datatables.net-bs\js\dataTables.bootstrap.min.js" />
-    <None Include="node_modules\datatables.net-bs\Readme.md" />
-    <None Include="node_modules\datatables.net\js\jquery.dataTables.js" />
-    <None Include="node_modules\datatables.net\js\jquery.dataTables.min.js" />
-    <None Include="node_modules\datatables.net\Readme.md" />
-    <None Include="node_modules\jquery\dist\jquery.js" />
-    <None Include="node_modules\jquery\dist\jquery.min.js" />
-    <None Include="node_modules\jquery\dist\jquery.min.map" />
-    <None Include="node_modules\jquery\dist\jquery.slim.js" />
-    <None Include="node_modules\jquery\dist\jquery.slim.min.js" />
-    <None Include="node_modules\jquery\dist\jquery.slim.min.map" />
-    <None Include="node_modules\jquery\external\sizzle\dist\sizzle.js" />
-    <None Include="node_modules\jquery\external\sizzle\dist\sizzle.min.js" />
-    <None Include="node_modules\jquery\external\sizzle\dist\sizzle.min.map" />
-    <None Include="node_modules\jquery\README.md" />
-    <None Include="node_modules\jquery\src\ajax.js" />
-    <None Include="node_modules\jquery\src\ajax\jsonp.js" />
-    <None Include="node_modules\jquery\src\ajax\load.js" />
-    <None Include="node_modules\jquery\src\ajax\script.js" />
-    <None Include="node_modules\jquery\src\ajax\var\location.js" />
-    <None Include="node_modules\jquery\src\ajax\var\nonce.js" />
-    <None Include="node_modules\jquery\src\ajax\var\rquery.js" />
-    <None Include="node_modules\jquery\src\ajax\xhr.js" />
-    <None Include="node_modules\jquery\src\attributes.js" />
-    <None Include="node_modules\jquery\src\attributes\attr.js" />
-    <None Include="node_modules\jquery\src\attributes\classes.js" />
-    <None Include="node_modules\jquery\src\attributes\prop.js" />
-    <None Include="node_modules\jquery\src\attributes\support.js" />
-    <None Include="node_modules\jquery\src\attributes\val.js" />
-    <None Include="node_modules\jquery\src\callbacks.js" />
-    <None Include="node_modules\jquery\src\core.js" />
-    <None Include="node_modules\jquery\src\core\access.js" />
-    <None Include="node_modules\jquery\src\core\camelCase.js" />
-    <None Include="node_modules\jquery\src\core\DOMEval.js" />
-    <None Include="node_modules\jquery\src\core\init.js" />
-    <None Include="node_modules\jquery\src\core\isAttached.js" />
-    <None Include="node_modules\jquery\src\core\nodeName.js" />
-    <None Include="node_modules\jquery\src\core\parseHTML.js" />
-    <None Include="node_modules\jquery\src\core\parseXML.js" />
-    <None Include="node_modules\jquery\src\core\ready-no-deferred.js" />
-    <None Include="node_modules\jquery\src\core\ready.js" />
-    <None Include="node_modules\jquery\src\core\readyException.js" />
-    <None Include="node_modules\jquery\src\core\stripAndCollapse.js" />
-    <None Include="node_modules\jquery\src\core\support.js" />
-    <None Include="node_modules\jquery\src\core\toType.js" />
-    <None Include="node_modules\jquery\src\core\var\rsingleTag.js" />
-    <None Include="node_modules\jquery\src\css.js" />
-    <None Include="node_modules\jquery\src\css\addGetHookIf.js" />
-    <None Include="node_modules\jquery\src\css\adjustCSS.js" />
-    <None Include="node_modules\jquery\src\css\curCSS.js" />
-    <None Include="node_modules\jquery\src\css\finalPropName.js" />
-    <None Include="node_modules\jquery\src\css\hiddenVisibleSelectors.js" />
-    <None Include="node_modules\jquery\src\css\showHide.js" />
-    <None Include="node_modules\jquery\src\css\support.js" />
-    <None Include="node_modules\jquery\src\css\var\cssExpand.js" />
-    <None Include="node_modules\jquery\src\css\var\getStyles.js" />
-    <None Include="node_modules\jquery\src\css\var\isHiddenWithinTree.js" />
-    <None Include="node_modules\jquery\src\css\var\rboxStyle.js" />
-    <None Include="node_modules\jquery\src\css\var\rnumnonpx.js" />
-    <None Include="node_modules\jquery\src\css\var\swap.js" />
-    <None Include="node_modules\jquery\src\data.js" />
-    <None Include="node_modules\jquery\src\data\Data.js" />
-    <None Include="node_modules\jquery\src\data\var\acceptData.js" />
-    <None Include="node_modules\jquery\src\data\var\dataPriv.js" />
-    <None Include="node_modules\jquery\src\data\var\dataUser.js" />
-    <None Include="node_modules\jquery\src\deferred.js" />
-    <None Include="node_modules\jquery\src\deferred\exceptionHook.js" />
-    <None Include="node_modules\jquery\src\deprecated.js" />
-    <None Include="node_modules\jquery\src\deprecated\ajax-event-alias.js" />
-    <None Include="node_modules\jquery\src\deprecated\event.js" />
-    <None Include="node_modules\jquery\src\dimensions.js" />
-    <None Include="node_modules\jquery\src\effects.js" />
-    <None Include="node_modules\jquery\src\effects\animatedSelector.js" />
-    <None Include="node_modules\jquery\src\effects\Tween.js" />
-    <None Include="node_modules\jquery\src\event.js" />
-    <None Include="node_modules\jquery\src\event\focusin.js" />
-    <None Include="node_modules\jquery\src\event\support.js" />
-    <None Include="node_modules\jquery\src\event\trigger.js" />
-    <None Include="node_modules\jquery\src\exports\amd.js" />
-    <None Include="node_modules\jquery\src\exports\global.js" />
-    <None Include="node_modules\jquery\src\jquery.js" />
-    <None Include="node_modules\jquery\src\manipulation.js" />
-    <None Include="node_modules\jquery\src\manipulation\buildFragment.js" />
-    <None Include="node_modules\jquery\src\manipulation\getAll.js" />
-    <None Include="node_modules\jquery\src\manipulation\setGlobalEval.js" />
-    <None Include="node_modules\jquery\src\manipulation\support.js" />
-    <None Include="node_modules\jquery\src\manipulation\var\rscriptType.js" />
-    <None Include="node_modules\jquery\src\manipulation\var\rtagName.js" />
-    <None Include="node_modules\jquery\src\manipulation\wrapMap.js" />
-    <None Include="node_modules\jquery\src\manipulation\_evalUrl.js" />
-    <None Include="node_modules\jquery\src\offset.js" />
-    <None Include="node_modules\jquery\src\queue.js" />
-    <None Include="node_modules\jquery\src\queue\delay.js" />
-    <None Include="node_modules\jquery\src\selector-native.js" />
-    <None Include="node_modules\jquery\src\selector-sizzle.js" />
-    <None Include="node_modules\jquery\src\selector.js" />
-    <None Include="node_modules\jquery\src\serialize.js" />
-    <None Include="node_modules\jquery\src\traversing.js" />
-    <None Include="node_modules\jquery\src\traversing\findFilter.js" />
-    <None Include="node_modules\jquery\src\traversing\var\dir.js" />
-    <None Include="node_modules\jquery\src\traversing\var\rneedsContext.js" />
-    <None Include="node_modules\jquery\src\traversing\var\siblings.js" />
-    <None Include="node_modules\jquery\src\var\arr.js" />
-    <None Include="node_modules\jquery\src\var\class2type.js" />
-    <None Include="node_modules\jquery\src\var\document.js" />
-    <None Include="node_modules\jquery\src\var\documentElement.js" />
-    <None Include="node_modules\jquery\src\var\flat.js" />
-    <None Include="node_modules\jquery\src\var\fnToString.js" />
-    <None Include="node_modules\jquery\src\var\getProto.js" />
-    <None Include="node_modules\jquery\src\var\hasOwn.js" />
-    <None Include="node_modules\jquery\src\var\indexOf.js" />
-    <None Include="node_modules\jquery\src\var\isFunction.js" />
-    <None Include="node_modules\jquery\src\var\isWindow.js" />
-    <None Include="node_modules\jquery\src\var\ObjectFunctionString.js" />
-    <None Include="node_modules\jquery\src\var\pnum.js" />
-    <None Include="node_modules\jquery\src\var\push.js" />
-    <None Include="node_modules\jquery\src\var\rcheckableType.js" />
-    <None Include="node_modules\jquery\src\var\rcssNum.js" />
-    <None Include="node_modules\jquery\src\var\rnothtmlwhite.js" />
-    <None Include="node_modules\jquery\src\var\slice.js" />
-    <None Include="node_modules\jquery\src\var\support.js" />
-    <None Include="node_modules\jquery\src\var\toString.js" />
-    <None Include="node_modules\jquery\src\wrap.js" />
-    <None Include="wwwroot\OdeToFood.png" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore" Version="2.1.7" />
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc" Version="2.1.3" />
-    <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.1.2" PrivateAssets="All" />
-    <PackageReference Include="Microsoft.AspNetCore.StaticFiles" Version="2.1.1" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="2.1.14" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="2.1.14" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="2.1.14">
+     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.1.0" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="3.1.0">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
-    <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="2.1.1" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.BrowserLink" Version="2.1.1" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.1.10" />
-    <PackageReference Include="OdeToCode.UseNodeModules" Version="1.0.6" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="3.1.0" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="3.1.0">
+      <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    </PackageReference>
+    <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="3.1.0" />
+    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="3.1.0" />
+    <PackageReference Include="OdeToCode.UseNodeModules" Version="3.0.0" />
   </ItemGroup>
 
   <ItemGroup>
