@@ -10,6 +10,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using OdeToFood.Data;
 
 namespace OdeToFood
@@ -31,7 +32,10 @@ namespace OdeToFood
         options.UseSqlServer(Configuration.GetConnectionString("OdeToFoodDb"));
       });
 
+      //Below is with an actual database
       services.AddScoped<IRestaurantData, SqlRestaurantData>();
+      //This is for IIS purposes
+      //services.AddScoped<IRestaurantData, InMemoryRestaurantData>();
 
       services.Configure<CookiePolicyOptions>(options =>
       {
@@ -40,12 +44,15 @@ namespace OdeToFood
         options.MinimumSameSitePolicy = SameSiteMode.None;
       });
 
-
-      services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
+      //aspnetcore 2.0
+      //services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
+      // aspnetcore30
+      services.AddRazorPages();
+      services.AddControllers();
     }
 
     // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-    public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+    public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
     {
       if (env.IsDevelopment())
       {
@@ -61,10 +68,17 @@ namespace OdeToFood
       app.Use(SayHelloMiddleware);
       app.UseHttpsRedirection();
       app.UseStaticFiles();
-      app.UseNodeModules(env);
+      app.UseNodeModules();
       app.UseCookiePolicy();
 
-      app.UseMvc();
+      //app.UseMvc();
+      // aspnetcore30
+      app.UseRouting();
+      app.UseEndpoints(e =>
+      {
+        e.MapRazorPages();
+        e.MapControllers();
+      });
     }
 
     private RequestDelegate SayHelloMiddleware(RequestDelegate next)
